Update issue reference in git guidelines
As a pull request contains all related commits there is no need to put a reference at the end of each commit message. The reference can be put only once in a pull request description so it appear in the issue history. To find a certain commit that resolves issue we can use this path: Issue -> Pull request -> Commit.